API/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
psycopg[binary]>=3.1
python-dotenv==1.0.1
